Move toy models to reaction_prediction/inference (and move their tests)
