+++ syntheseus/tests/search/conftest.py
@@ -18,7 +18,7 @@ from syntheseus.search.graph.molset import MolSetGraph
 from syntheseus.search.graph.route import SynthesisGraph
 from syntheseus.search.mol_inventory import BaseMolInventory, SmilesListInventory
 from syntheseus.search.reaction_models import BackwardReactionModel
-from syntheseus.search.reaction_models.toy import LinearMolecules, ListOfReactionsModel
+from syntheseus.search.reaction_models.toy import LinearMoleculesToyModel, ListOfReactionsToyModel
 
 
 @dataclass
@@ -33,18 +33,8 @@ class RetrosynthesisTask:
 
 
 @pytest.fixture
-def rxn_cs_from_cc() -> SingleProductReaction:
-    return SingleProductReaction(product=Molecule("CS"), reactants=Bag([Molecule("CC")]))
-
-
-@pytest.fixture
 def rxn_cs_from_co() -> SingleProductReaction:
     return SingleProductReaction(product=Molecule("CS"), reactants=Bag([Molecule("CO")]))
-
-
-@pytest.fixture
-def rxn_cocs_from_cocc(cocs_mol: Molecule) -> SingleProductReaction:
-    return SingleProductReaction(product=cocs_mol, reactants=Bag([Molecule("COCC")]))
 
 
 @pytest.fixture
@@ -104,7 +94,7 @@ def retrosynthesis_task1(
 
     return RetrosynthesisTask(
         target_mol=cocs_mol,
-        reaction_model=LinearMolecules(),
+        reaction_model=LinearMoleculesToyModel(),
         inventory=SmilesListInventory(["CO", "CS"]),
         known_routes={"min-cost": best_route, "other": other_route},
     )
@@ -177,7 +167,7 @@ def retrosynthesis_task2(
 
     return RetrosynthesisTask(
         target_mol=cocs_mol,
-        reaction_model=LinearMolecules(),
+        reaction_model=LinearMoleculesToyModel(),
         inventory=SmilesListInventory(["CO", "CC"]),
         known_routes=known_routes,
         incorrect_routes=incorrect_routes,
@@ -195,7 +185,9 @@ def retrosynthesis_task3(cocs_mol: Molecule) -> RetrosynthesisTask:
     CO -> CC
     """
     return RetrosynthesisTask(
-        target_mol=cocs_mol, reaction_model=LinearMolecules(), inventory=SmilesListInventory(["CC"])
+        target_mol=cocs_mol,
+        reaction_model=LinearMoleculesToyModel(),
+        inventory=SmilesListInventory(["CC"]),
     )
 
 
@@ -212,7 +204,7 @@ def retrosynthesis_task4(cocs_mol: Molecule) -> RetrosynthesisTask:
     """
     return RetrosynthesisTask(
         target_mol=cocs_mol,
-        reaction_model=LinearMolecules(allow_substitution=False),
+        reaction_model=LinearMoleculesToyModel(allow_substitution=False),
         inventory=SmilesListInventory(["C", "O", "S"]),
     )
 
@@ -225,7 +217,7 @@ def retrosynthesis_task5() -> RetrosynthesisTask:
     """
     return RetrosynthesisTask(
         target_mol=Molecule("CC", make_rdkit_mol=False),
-        reaction_model=LinearMolecules(allow_substitution=True),
+        reaction_model=LinearMoleculesToyModel(allow_substitution=True),
         inventory=SmilesListInventory(["O"]),
     )
 
@@ -250,7 +242,7 @@ def retrosynthesis_task6() -> RetrosynthesisTask:
     """
     return RetrosynthesisTask(
         target_mol=Molecule("CCCOC", make_rdkit_mol=False),
-        reaction_model=LinearMolecules(allow_substitution=True),
+        reaction_model=LinearMoleculesToyModel(allow_substitution=True),
         inventory=SmilesListInventory(["CCCO", "CC", "COC", "O"]),
     )
 
@@ -260,7 +252,7 @@ def rxn_model_for_minimal_graphs(
     rxn_cocs_from_co_cs: SingleProductReaction,
     rxn_co_from_cc: SingleProductReaction,
     rxn_cs_from_co: SingleProductReaction,
-) -> ListOfReactionsModel:
+) -> ListOfReactionsToyModel:
     """
     Return a reaction model to help build the minimal graphs.
     Contains the following reactions:
@@ -269,12 +261,14 @@ def rxn_model_for_minimal_graphs(
     CO -> CC
     CS -> CO
     """
-    return ListOfReactionsModel(reaction_list=[rxn_cocs_from_co_cs, rxn_co_from_cc, rxn_cs_from_co])
+    return ListOfReactionsToyModel(
+        reaction_list=[rxn_cocs_from_co_cs, rxn_co_from_cc, rxn_cs_from_co]
+    )
 
 
 @pytest.fixture
 def rxn_model_for_non_minimal_graphs(
-    rxn_model_for_minimal_graphs: ListOfReactionsModel,
+    rxn_model_for_minimal_graphs: ListOfReactionsToyModel,
     rxn_cocs_from_cocc: SingleProductReaction,
     rxn_cocc_from_co_cc: SingleProductReaction,
 ) -> BackwardReactionModel:
